requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
atlassian-python-api==3.40.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.13
Deprecated==1.2.14
exceptiongroup==1.1.2
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
html2text==2020.1.16
httpcore==0.17.3
httpx==0.24.1
idna==3.4
langchain==0.0.247
langsmith==0.0.15
llama-index==0.7.14
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pydantic==1.10.12
pypdf==3.14.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
slackclient==2.9.4
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
